+++ app/models/item.py
@@ -4,19 +4,6 @@ from sqlalchemy.orm import relationship
 # import db file from my folder config
 from app.config.db import Base
 
-base2 = Base
-
-
-class User(Base):
-    __tablename__ = "users"
-
-    id = Column(Integer, primary_key=True, index=True)
-    email = Column(String, unique=True, index=True)
-    hashed_password = Column(String)
-    is_active = Column(Boolean, default=True)
-
-    items = relationship("Item", back_populates="owner")
-
 class Item(Base):
     __tablename__ = "items"
